+++ include/circle/usb/dwhcidevice.h
@@ -2,7 +2,7 @@
 // dwhcidevice.h
 //
 // Circle - A C++ bare metal environment for Raspberry Pi
-// Copyright (C) 2014-2020  R. Stange <rsta2@o2online.de>
+// Copyright (C) 2014-2022  R. Stange <rsta2@o2online.de>
 // 
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -28,10 +28,12 @@
 #include <circle/usb/dwhcirootport.h>
 #include <circle/usb/dwhcixferstagedata.h>
 #include <circle/usb/dwhcixactqueue.h>
+#include <circle/usb/dwhcicompletionqueue.h>
 #include <circle/usb/dwhciregister.h>
 #include <circle/usb/dwhci.h>
 #include <circle/usb/usb.h>
 #include <circle/spinlock.h>
+#include <circle/mphi.h>
 #include <circle/sysconfig.h>
 #include <circle/types.h>
 
@@ -100,6 +102,11 @@ private:
 	void InterruptHandler (void);
 	static void InterruptStub (void *pParam);
 
+#ifdef USE_USB_FIQ
+	void InterruptHandler2 (void);
+	static void InterruptStub2 (void *pParam);
+#endif
+
 #ifndef USE_USB_SOF_INTR
 	void TimerHandler (CDWHCITransferStageData *pStageData);
 	static void TimerStub (TKernelTimerHandle hTimer, void *pParam, void *pContext);
@@ -115,7 +122,9 @@ private:
 			    u32		    nMask,
 			    boolean	    bWaitUntilSet,
 			    unsigned	    nMsTimeout);
-	
+
+	void LogTransactionFailed (u32 nStatus);
+
 #ifndef NDEBUG
 	void DumpRegister (const char *pName, u32 nAddress);
 	void DumpStatus (unsigned nChannel = 0);
@@ -143,6 +152,12 @@ private:
 
 	CDWHCIRootPort m_RootPort;
 	volatile boolean m_bRootPortEnabled;
+
+#ifdef USE_USB_FIQ
+	volatile int m_nPortStatusChanged;
+	CDWHCICompletionQueue m_CompletionQueue;
+	CMPHIDevice m_MPHI;
+#endif
 
 	volatile boolean m_bShutdown;			// USB driver will shutdown
 };