+++ lib/usb/dwhcidevice.cpp
@@ -6,7 +6,7 @@
 //	no ISO transfers
 //
 // Circle - A C++ bare metal environment for Raspberry Pi
-// Copyright (C) 2014-2021  R. Stange <rsta2@o2online.de>
+// Copyright (C) 2014-2022  R. Stange <rsta2@o2online.de>
 // 
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -32,6 +32,7 @@
 #include <circle/logger.h>
 #include <circle/koptions.h>
 #include <circle/sysconfig.h>
+#include <circle/atomic.h>
 #include <circle/debug.h>
 #include <assert.h>
 
@@ -42,6 +43,12 @@
 	#define DWC_CFG_HOST_RX_FIFO_SIZE	1024	// number of 32 bit words
 	#define DWC_CFG_HOST_NPER_TX_FIFO_SIZE	1024	// number of 32 bit words
 	#define DWC_CFG_HOST_PER_TX_FIFO_SIZE	1024	// number of 32 bit words
+
+#ifdef USE_USB_FIQ
+	#define MAX_TARGET_LEVEL	FIQ_LEVEL
+#else
+	#define MAX_TARGET_LEVEL	IRQ_LEVEL
+#endif
 
 enum TStageState
 {
@@ -59,7 +66,7 @@ enum TStageSubState
 	StageSubStateUnknown
 };
 
-static const char FromDWHCI[] = "dwhci";
+LOGMODULE ("dwhci");
 
 CDWHCIDevice::CDWHCIDevice (CInterruptSystem *pInterruptSystem, CTimer *pTimer, boolean bPlugAndPlay)
 :	CUSBHostController (bPlugAndPlay),
@@ -67,10 +74,20 @@ CDWHCIDevice::CDWHCIDevice (CInterruptSystem *pInterruptSystem, CTimer *pTimer, 
 	m_pTimer (pTimer),
 	m_nChannels (0),
 	m_nChannelAllocated (0),
+	m_ChannelSpinLock (MAX_TARGET_LEVEL),
+#ifdef USE_USB_SOF_INTR
+	m_TransactionQueue (DWHCI_MAX_CHANNELS, MAX_TARGET_LEVEL),
+#endif
+	m_IntMaskSpinLock (MAX_TARGET_LEVEL),
 	m_nWaitBlockAllocated (0),
 	m_WaitBlockSpinLock (TASK_LEVEL),
 	m_RootPort (this),
 	m_bRootPortEnabled (FALSE),
+#ifdef USE_USB_FIQ
+	m_nPortStatusChanged (0),
+	m_CompletionQueue (DWHCI_MAX_CHANNELS*2),
+	m_MPHI (pInterruptSystem),
+#endif
 	m_bShutdown (FALSE)
 {
 	assert (m_pInterruptSystem != 0);
@@ -95,7 +112,12 @@ CDWHCIDevice::~CDWHCIDevice (void)
 	m_pTimer->MsDelay (200);	// wait for completion of all transactions
 
 	assert (m_pInterruptSystem != 0);
+#ifndef USE_USB_FIQ
 	m_pInterruptSystem->DisconnectIRQ (ARM_IRQ_USB);
+#else
+	m_pInterruptSystem->DisconnectFIQ ();
+	m_MPHI.DisconnectHandler ();
+#endif
 
 	Reset ();
 
@@ -114,18 +136,16 @@ boolean CDWHCIDevice::Initialize (boolean bScanDevices)
 #ifndef USE_USB_SOF_INTR
 	if (IsPlugAndPlay ())
 	{
-		CLogger::Get ()->Write (FromDWHCI, LogWarning,
-					"Using plug-and-play without USE_USB_SOF_INTR "
-					"is not recommended");
+		LOGWARN ("Using plug-and-play without USE_USB_SOF_INTR is not recommended");
 	}
 #endif
 
 	// init class-specific allocators in USB library
-	INIT_PROTECTED_CLASS_ALLOCATOR (CUSBRequest, DWHCI_MAX_CHANNELS*2, IRQ_LEVEL);
-	INIT_PROTECTED_CLASS_ALLOCATOR (CDWHCITransferStageData, DWHCI_MAX_CHANNELS, IRQ_LEVEL);
-	INIT_PROTECTED_CLASS_ALLOCATOR (CDWHCIFrameSchedulerNonPeriodic, DWHCI_MAX_CHANNELS, IRQ_LEVEL);
-	INIT_PROTECTED_CLASS_ALLOCATOR (CDWHCIFrameSchedulerPeriodic, DWHCI_MAX_CHANNELS, IRQ_LEVEL);
-	INIT_PROTECTED_CLASS_ALLOCATOR (CDWHCIFrameSchedulerNoSplit, DWHCI_MAX_CHANNELS, IRQ_LEVEL);
+	INIT_PROTECTED_CLASS_ALLOCATOR (CUSBRequest, DWHCI_MAX_CHANNELS*2, MAX_TARGET_LEVEL);
+	INIT_PROTECTED_CLASS_ALLOCATOR (CDWHCITransferStageData, DWHCI_MAX_CHANNELS, MAX_TARGET_LEVEL);
+	INIT_PROTECTED_CLASS_ALLOCATOR (CDWHCIFrameSchedulerNonPeriodic, DWHCI_MAX_CHANNELS, MAX_TARGET_LEVEL);
+	INIT_PROTECTED_CLASS_ALLOCATOR (CDWHCIFrameSchedulerPeriodic, DWHCI_MAX_CHANNELS, MAX_TARGET_LEVEL);
+	INIT_PROTECTED_CLASS_ALLOCATOR (CDWHCIFrameSchedulerNoSplit, DWHCI_MAX_CHANNELS, MAX_TARGET_LEVEL);
 
 	PeripheralEntry ();
 
@@ -135,13 +155,13 @@ boolean CDWHCIDevice::Initialize (boolean bScanDevices)
 	CDWHCIRegister VendorId (DWHCI_CORE_VENDOR_ID);
 	if (VendorId.Read () != 0x4F54280A)
 	{
-		CLogger::Get ()->Write (FromDWHCI, LogError, "Unknown vendor 0x%0X", VendorId.Get ());
+		LOGERR ("Unknown vendor 0x%0X", VendorId.Get ());
 		return FALSE;
 	}
 
 	if (!PowerOn ())
 	{
-		CLogger::Get ()->Write (FromDWHCI, LogError, "Cannot power on");
+		LOGERR ("Cannot power on");
 		return FALSE;
 	}
 	
@@ -152,11 +172,16 @@ boolean CDWHCIDevice::Initialize (boolean bScanDevices)
 	AHBConfig.Write ();
 	
 	assert (m_pInterruptSystem != 0);
+#ifndef USE_USB_FIQ
 	m_pInterruptSystem->ConnectIRQ (ARM_IRQ_USB, InterruptStub, this);
+#else
+	m_MPHI.ConnectHandler (InterruptStub2, this);
+	m_pInterruptSystem->ConnectFIQ (ARM_FIQ_USB, InterruptStub, this);
+#endif
 
 	if (!InitCore ())
 	{
-		CLogger::Get ()->Write (FromDWHCI, LogError, "Cannot initialize core");
+		LOGERR ("Cannot initialize core");
 		return FALSE;
 	}
 	
@@ -164,7 +189,7 @@ boolean CDWHCIDevice::Initialize (boolean bScanDevices)
 	
 	if (!InitHost ())
 	{
-		CLogger::Get ()->Write (FromDWHCI, LogError, "Cannot initialize host");
+		LOGERR ("Cannot initialize host");
 		return FALSE;
 	}
 
@@ -191,14 +216,12 @@ void CDWHCIDevice::ReScanDevices (void)
 
 			if (!m_RootPort.Initialize ())
 			{
-				CLogger::Get ()->Write (FromDWHCI, LogWarning,
-							"Cannot initialize root port");
+				LOGERR ("Cannot initialize root port");
 			}
 		}
 		else
 		{
-			CLogger::Get ()->Write (FromDWHCI, LogWarning,
-						"No device connected to root port");
+			LOGWARN ("No device connected to root port");
 		}
 	}
 	else
@@ -390,7 +413,7 @@ boolean CDWHCIDevice::InitCore (void)
 
 	if (!Reset ())
 	{
-		CLogger::Get ()->Write (FromDWHCI, LogError, "Reset failed");
+		LOGERR ("Reset failed");
 		return FALSE;
 	}
 
@@ -1011,7 +1034,11 @@ void CDWHCIDevice::ChannelInterruptHandler (unsigned nChannel)
 
 		FreeChannel (nChannel);
 
+#ifndef USE_USB_FIQ
 		pURB->CallCompletionRoutine ();
+#else
+		m_CompletionQueue.Enqueue (pURB);
+#endif
 
 		return;
 	}
@@ -1081,8 +1108,7 @@ void CDWHCIDevice::ChannelInterruptHandler (unsigned nChannel)
 		}
 		else if (nStatus & DWHCI_HOST_CHAN_INT_ERROR_MASK)
 		{
-			CLogger::Get ()->Write (FromDWHCI, LogError,
-						"Transaction failed (status 0x%X)", nStatus);
+			LogTransactionFailed (nStatus);
 
 			pURB->SetStatus (0);
 			pURB->SetUSBError (pStageData->GetUSBError ());
@@ -1131,7 +1157,11 @@ void CDWHCIDevice::ChannelInterruptHandler (unsigned nChannel)
 
 		FreeChannel (nChannel);
 
+#ifndef USE_USB_FIQ
 		pURB->CallCompletionRoutine ();
+#else
+		m_CompletionQueue.Enqueue (pURB);
+#endif
 		break;
 
 	case StageStateStartSplit:
@@ -1140,8 +1170,7 @@ void CDWHCIDevice::ChannelInterruptHandler (unsigned nChannel)
 		    || (nStatus & DWHCI_HOST_CHAN_INT_NAK)
 		    || (nStatus & DWHCI_HOST_CHAN_INT_NYET))
 		{
-			CLogger::Get ()->Write (FromDWHCI, LogError,
-						"Transaction failed (status 0x%X)", nStatus);
+			LogTransactionFailed (nStatus);
 
 			pURB->SetStatus (0);
 			pURB->SetUSBError (pStageData->GetUSBError ());
@@ -1153,7 +1182,11 @@ void CDWHCIDevice::ChannelInterruptHandler (unsigned nChannel)
 
 			FreeChannel (nChannel);
 
+#ifndef USE_USB_FIQ
 			pURB->CallCompletionRoutine ();
+#else
+			m_CompletionQueue.Enqueue (pURB);
+#endif
 			break;
 		}
 
@@ -1181,8 +1214,7 @@ void CDWHCIDevice::ChannelInterruptHandler (unsigned nChannel)
 		nStatus = pStageData->GetTransactionStatus ();
 		if (nStatus & DWHCI_HOST_CHAN_INT_ERROR_MASK)
 		{
-			CLogger::Get ()->Write (FromDWHCI, LogError,
-						"Transaction failed (status 0x%X)", nStatus);
+			LogTransactionFailed (nStatus);
 
 			pURB->SetStatus (0);
 			pURB->SetUSBError (pStageData->GetUSBError ());
@@ -1194,7 +1226,11 @@ void CDWHCIDevice::ChannelInterruptHandler (unsigned nChannel)
 
 			FreeChannel (nChannel);
 
+#ifndef USE_USB_FIQ
 			pURB->CallCompletionRoutine ();
+#else
+			m_CompletionQueue.Enqueue (pURB);
+#endif
 			break;
 		}
 		
@@ -1228,7 +1264,11 @@ void CDWHCIDevice::ChannelInterruptHandler (unsigned nChannel)
 
 				FreeChannel (nChannel);
 
+#ifndef USE_USB_FIQ
 				pURB->CallCompletionRoutine ();
+#else
+				m_CompletionQueue.Enqueue (pURB);
+#endif
 				break;
 			}
 
@@ -1261,7 +1301,11 @@ void CDWHCIDevice::ChannelInterruptHandler (unsigned nChannel)
 
 					FreeChannel (nChannel);
 
+#ifndef USE_USB_FIQ
 					pURB->CallCompletionRoutine ();
+#else
+					m_CompletionQueue.Enqueue (pURB);
+#endif
 				}
 				else
 				{
@@ -1296,7 +1340,11 @@ void CDWHCIDevice::ChannelInterruptHandler (unsigned nChannel)
 
 		FreeChannel (nChannel);
 
+#ifndef USE_USB_FIQ
 		pURB->CallCompletionRoutine ();
+#else
+		m_CompletionQueue.Enqueue (pURB);
+#endif
 		break;
 
 	default:
@@ -1321,10 +1369,8 @@ void CDWHCIDevice::SOFInterruptHandler (void)
 	while ((pStageData = m_TransactionQueue.Dequeue (usFrameNumber)) != 0)
 	{
 		unsigned nChannel = AllocateChannel ();
-		if (nChannel >= m_nChannels)
-		{
-			CLogger::Get ()->Write (FromDWHCI, LogPanic, "Too many parallel USB transactions");
-		}
+		assert (nChannel < m_nChannels);	// too many parallel transactions otherwise
+
 		pStageData->SetChannelNumber (nChannel);
 
 		assert (m_pStageData[nChannel] == 0);
@@ -1340,21 +1386,28 @@ void CDWHCIDevice::SOFInterruptHandler (void)
 
 void CDWHCIDevice::InterruptHandler (void)
 {
-#ifndef NDEBUG
-	//debug_click ();
-#endif
-
 	PeripheralEntry ();
 
 	CDWHCIRegister IntStatus (DWHCI_CORE_INT_STAT);
 	IntStatus.Read ();
+
+#ifdef USE_USB_SOF_INTR
+	if (IntStatus.Get () & DWHCI_CORE_INT_STAT_SOF_INTR)
+	{
+#ifndef NDEBUG
+		//debug_click (DEBUG_CLICK_LEFT);
+#endif
+
+		SOFInterruptHandler ();
+	}
+#endif
 
 	if (IntStatus.Get () & DWHCI_CORE_INT_STAT_HC_INTR)
 	{
 		CDWHCIRegister AllChanInterrupt (DWHCI_HOST_ALLCHAN_INT);
 		AllChanInterrupt.Read ();
 		AllChanInterrupt.Write ();
-		
+
 		unsigned nChannelMask = 1;
 		for (unsigned nChannel = 0; nChannel < m_nChannels; nChannel++)
 		{
@@ -1365,17 +1418,10 @@ void CDWHCIDevice::InterruptHandler (void)
 				
 				ChannelInterruptHandler (nChannel);
 			}
-			
+
 			nChannelMask <<= 1;
 		}
 	}
-
-#ifdef USE_USB_SOF_INTR
-	if (IntStatus.Get () & DWHCI_CORE_INT_STAT_SOF_INTR)
-	{
-		SOFInterruptHandler ();
-	}
-#endif
 
 	if (IsPlugAndPlay ())
 	{
@@ -1384,12 +1430,15 @@ void CDWHCIDevice::InterruptHandler (void)
 			CDWHCIRegister HostPort (DWHCI_HOST_PORT);
 			HostPort.Read ();
 
-			//CLogger::Get ()->Write (FromDWHCI, LogDebug,
-			//			"Port interrupt (status 0x%08X)", HostPort.Get ());
+			//LOGDBG ("Port interrupt (status 0x%08X)", HostPort.Get ());
 
 			if (HostPort.Get () & DWHCI_HOST_PORT_CONNECT_CHANGED)
 			{
+#ifndef USE_USB_FIQ
 				PortStatusChanged (&m_RootPort);
+#else
+				AtomicSet (&m_nPortStatusChanged, 1);
+#endif
 			}
 
 			HostPort.And (~DWHCI_HOST_PORT_ENABLE);
@@ -1401,15 +1450,27 @@ void CDWHCIDevice::InterruptHandler (void)
 
 		if (IntStatus.Get () & DWHCI_CORE_INT_MASK_DISCONNECT)
 		{
-			//CLogger::Get ()->Write (FromDWHCI, LogDebug, "Disconnect interrupt");
+			//LOGDBG ("Disconnect interrupt");
 
+#ifndef USE_USB_FIQ
 			PortStatusChanged (&m_RootPort);
+#else
+			AtomicSet (&m_nPortStatusChanged, 1);
+#endif
 		}
 	}
 
 	IntStatus.Write ();
 
 	PeripheralExit ();
+
+#ifdef USE_USB_FIQ
+	if (   !m_CompletionQueue.IsEmpty ()
+	    || AtomicGet (&m_nPortStatusChanged))
+	{
+		m_MPHI.TriggerIRQ ();
+	}
+#endif
 }
 
 void CDWHCIDevice::InterruptStub (void *pParam)
@@ -1480,6 +1541,36 @@ void CDWHCIDevice::TimerStub (TKernelTimerHandle /* hTimer */, void *pParam, voi
 	assert (pParam != 0);
 	
 	pThis->TimerHandler (pStageData);
+}
+
+#endif
+
+#ifdef USE_USB_FIQ
+
+void CDWHCIDevice::InterruptHandler2 (void)
+{
+#ifndef NDEBUG
+	//debug_click (DEBUG_CLICK_RIGHT);
+#endif
+
+	CUSBRequest *pURB;
+	while ((pURB = m_CompletionQueue.Dequeue ()) != 0)
+	{
+		pURB->CallCompletionRoutine ();
+	}
+
+	while (AtomicExchange (&m_nPortStatusChanged, 0))
+	{
+		PortStatusChanged (&m_RootPort);
+	}
+}
+
+void CDWHCIDevice::InterruptStub2 (void *pParam)
+{
+	CDWHCIDevice *pThis = static_cast <CDWHCIDevice *> (pParam);
+	assert (pThis != 0);
+
+	pThis->InterruptHandler2 ();
 }
 
 #endif
@@ -1574,7 +1665,7 @@ boolean CDWHCIDevice::WaitForBit (CDWHCIRegister *pRegister,
 
 		if (--nMsTimeout == 0)
 		{
-			//CLogger::Get ()->Write (FromDWHCI, LogWarning, "Timeout");
+			//LOGWARN ("Timeout");
 #ifndef NDEBUG
 			//pRegister->Dump ();
 #endif
@@ -1585,6 +1676,14 @@ boolean CDWHCIDevice::WaitForBit (CDWHCIRegister *pRegister,
 	return TRUE;
 }
 
+void CDWHCIDevice::LogTransactionFailed (u32 nStatus)
+{
+	if (CurrentExecutionLevel () < FIQ_LEVEL)
+	{
+		LOGWARN ("Transaction failed (status 0x%X)", nStatus);
+	}
+}
+
 #ifndef NDEBUG
 
 void CDWHCIDevice::DumpRegister (const char *pName, u32 nAddress)
@@ -1593,7 +1692,7 @@ void CDWHCIDevice::DumpRegister (const char *pName, u32 nAddress)
 
 	DataMemBarrier ();
 
-	CLogger::Get ()->Write (FromDWHCI, LogDebug, "0x%08X %s", Register.Read (), pName);
+	LOGDBG ("0x%08X %s", Register.Read (), pName);
 }
 
 void CDWHCIDevice::DumpStatus (unsigned nChannel)